Bug fixes
Fixed bugs where some of the boxes' corner colliders were not working properly with the MagnetGun. The BlueStarUI also had some bugs, fixed those as well, as I do. Also added a timer to switch from the last scene to the Main menu (going to change it to credits eventually, issues)